Got rid of classes no longer needed
Deleted the transition async task no longer needed, because one task
handles all transitions now.  Also got rid of the old async tasks for
running the counters, and their respective listeners, because the
counters run on a thread and post runnables to the UI thread now.  That
was an old change.