Added dynamic difficulty switching.
Now if the Shared Prefs difficulty setting is changed, the accelerator is set to
what is would be had the user been playing the entired time on the difficulty
aetting.  Now the reading of incoming intents needs to be worked on in Counter
Activity, the level is increasing if the user navigates away from the activity
to change the difficulty or view the game stats.  It should stay the same.  Going
to mack a Debug branch for this and math bugs.