Worked on Intents and SharedPrefs
Using startActivityForResult to launch FadeCounter, so Activities do not pile
up in the back stack.  Had to change a lot with resetting the values for a new
level after FadeCounter because we are reusing the same CounterActivity.  The
level does not increment after changing the difficulty of checking the game stats.
So far things seem to be working as they should with difficulty changed and
moving onto new levels.

* Need to debug the math in calculating accuracy

* Need to move the screen display methods to a new class, and hook up
  FadeCounter to them