Debugged the math in CounterActivity
Now the game grabs the accelerated count double, and rounds it to 2 decimal
places in onCounterStopped.  The rounded double is converted to a string to
display in the Counter TextView.  Also moved when OnCounterStopped is called
to whenever the background thread is interupted, instead of right when the
stop button is hit, because the loop usually gets in another incrementation
of advancing the counter after the stop button is hit.  This way all of the
math values match.

* Need to write a class for updating the game info displays (score, level,
  etc) that both Counter Activities can use.