Made changes in game info stats more obvious
Added blinking and color change to textviews that display relevant game stats changes
in ResetTurnAsync. Now the points added flash in green, lives gained flashes in green
and lifes lost flashes in red in CounterActivity. In FadeCounter, the lives only flash
if the accuracy is 100%. No score is awarded in FadeCounter, so nothing happens in the
score textview.