Changes to acceleration algorithm
Changed the Handler method post() to postAtTime(), and the mElapsedTimeMillis in
Counter Activity is passed into the uptimeMillis parameter.  Now the accelerated
count reliably posts at every hundreth of the counter (until the acceleration gets
too fast).  Before this change, the post calls were bottlenecking, so sometimes
the call would get skipped, and other times it would poat the same number several
times in a row in one iteration of the loop.  Need to work on a way to reduce the
frequency of posts when the acceleration gets too fast to post on every hundreth.