Overall changes and tweaks
Went through the code after the acceleration algorithm was working well.  Styles the dialogs,
added incrementing targets to the FadeCounter Activity.  Added and styled the ActionBar.  The
could still use a final pass through to get rid of uneeded classes, and methods.  It is working
pretty well right now besides that!