Bug fixes and packages
Fixed a bug in the ResetTurnAsync that was causing a brief dark blip when the counter
was fading away between turns. Also fixed a bug in the updating of the level in the
DB. Sorted the classes into packages.  Changed the way the accuracy is calculated,
so 100% only happens if the counter is dead on.  Now there are two bonus lifes if
you hit 100% accuracy, and one bonus life for 99%. Also changed the speeds to faster
settings.