Refactored FadeOutCounter
Same typre of changes as on Counter Activity on the last commit. Also refactored
the RestTurnAsync.  Wrote a method for the game info blinks after the turn. Still
have a problem with the counter hitting the max value and not resetting because
of the refactoring.  Need to trigger onCounterStopped from the background thread
if the max value is reached.