Tweaked the way score is calculated
The score is now caluclated within the margin of +2 or -2 of the target.
As the target gets higher and the accelerator quicker, this ensures that
the game really gets harder.  Before, calculating the accuracy on the
whole range from 0 to the target produced too wide of a margin, and in
some respects the game became easier as you went on.