SharedPrefs tracks if db exists
SharedPrefs now tracks if the db exists.  if it does not, (very first time
the game opens), it prevents the user from being able to view the game stats
.  Once the db does exist, the user can view the game stats at any time,
even on a new game before a turn has been played (they will only see previous games data until the play a turn in the new game).