Handling Back Button Events
In FadeCounter, a back button event sends an intent back to Counter
Activity, and advances the user to the next level, as if they didnt
score any bonus lives.  In Counter Activity, the onBackPressed()
sets the state.isFirstTurnInNewGame to true, so that the db adds
a new row.  The Activity was destroyed, but the AppState was still
alive, so the isFirstTurn was not reset as it would be if the
whole app was shut down.