Added Roboto as asset
The font was changing on Samsung devices if there was a default
system wide font set on the device.  Added Roboto regular as an
asset and set the font programmatically.  Also changed all textSize
declarations to dp instead of sp to prevent default device font
size settings from changing the layout.